+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Dates are in YYYY-MM-DD format.
 - Adds `SBuildrBuilder` for building projects using the SBuildr build system.
 - Renames `Generator` and associated classes/files to `Backend`.
 - Pulls in most functions from `cli` into `Project` to enable more powerful scripting.
+- `Project`'s `configure()` function now pickles and writes the Project into the build directory. `Project.load()` can be used to retrieve it.
+- `Project`'s `configure()` is now `freeze()`
+- Removes lazy header scanning - file manager now scans source files as they are added.
 
 ## v0.4.1 (2019-07-11)
 - Changes generator to favor false positives (longer builds) for `needs_configure()` rather than false negatives (broken builds).